fix sermon boundary issues
